refactor(repl): regroup into folders
