+++ packages/docs/src/repl/bundler/repl-dependencies-main.ts
