+++ packages/docs/src/repl/bundler/repl-dependencies.ts
