+++ packages/docs/src/repl/repl-ui/repl-detail-panel.tsx
